refactor(ui): decompose dashboard feature pages and hooks
Split DashboardPage.jsx (HIGH RISK grace/appear/sticky-latch state
machine into useDashboardPageState, empty-state content into
DashboardPageEmptyStates), ProjectsPage.jsx (data pipeline into
useProjectsPageData, pull-to-local into usePullToLocal),
AccumulatedOverviewPanel.jsx and RunOverviewPanel.jsx (hero sections,
computations, and a shared SeverityBadgeRow), DimensionGaugeCard.jsx
(coverage math), usePublish.js (queries/polling/optimistic-cache),
useDashboard.js (invalidation), useSharedProjects.js (coalesced
refresh), and RunHistoryPanel.jsx (chart defs).

Removed 3 now-compliant entries from the size grandfather list
(RunHistoryPanel.jsx, RunOverviewPanel.jsx, usePublish.js) and
regenerated the Python size baseline (108 -> 106 entries).
