+++ src/quodeq/ui/tools/size_grandfather.mjs
@@ -49,16 +49,13 @@ export const SIZE_GRANDFATHER = [
   "src/features/dashboard/components/OverviewSkeleton.test.jsx",
   "src/features/dashboard/components/ProjectsPage.jsx",
   "src/features/dashboard/components/ProjectsPage.test.jsx",
-  "src/features/dashboard/components/RunHistoryPanel.jsx",
   "src/features/dashboard/components/RunHistoryPanel.test.jsx",
   "src/features/dashboard/components/RunNavigator.jsx",
-  "src/features/dashboard/components/RunOverviewPanel.jsx",
   "src/features/dashboard/components/RunOverviewPanel.test.jsx",
   "src/features/dashboard/hooks/useDashboard.js",
   "src/features/dashboard/hooks/useDashboard.test.jsx",
   "src/features/dashboard/hooks/usePrefetchAdjacentRuns.test.jsx",
   "src/features/dashboard/hooks/usePrefetchRun.test.jsx",
-  "src/features/dashboard/hooks/usePublish.js",
   "src/features/dashboard/hooks/usePublish.test.jsx",
   "src/features/dashboard/hooks/useSharedProjects.js",
   "src/features/dashboard/hooks/useSharedProjects.test.jsx",
